requirements.txt
py-cord>=2.6.1
sqlalchemy>=2.0.3
python-dotenv>=1.0.1
pendulum>=3.0.0
aiosqlite>=0.21.0